decided no no auto type conversion. need type convertor functions

+++ clib/blah.c
@@ -1,5 +1,5 @@
 
-double fib_recursive(int n) {
+double fib_recursive(double n) {
     if (n < 2)
         return 1;
     return fib_recursive(n-1)+fib_recursive(n-2);
